tests/wpt/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/navigation_unload_data_url.html

<!doctype html>
<title>Same-origin navigation started from unload handler</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src="navigation_unload_data_url-1.html"></iframe>
<script>
var t = async_test();
onload = t.step_func(function() {
  var iframe = document.getElementsByTagName("iframe")[0];
  var new_src = "data:text/html,load<script>parent.postMessage('pass', '*')<\/script>";
  iframe.src = new_src;
});

onmessage = t.step_func(function(e) {
  assert_equals(e.data, "pass");
  t.done();
});

</script>
